requirements.txt
python-dotenv==0.19.0
livekit-server-sdk==1.0.6
google-api-python-client==2.15.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
openai==1.3.7
elevenlabs==0.3.0
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pydantic[email]==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 